Create null class with timestamp
