+++ src/NullObject.php
@@ -4,60 +4,17 @@ declare(strict_types=1);
 
 namespace Koriym\NullObject;
 
-use Koriym\NullObject\Exception\FileNotWritable;
-use Koriym\NullObject\Exception\LogicException;
-use ReflectionAttribute;
-use ReflectionClass;
-use ReflectionMethod;
-
-use function array_slice;
-use function assert;
-use function class_exists;
-use function dirname;
-use function file;
-use function file_put_contents;
-use function filemtime;
-use function hash;
-use function implode;
-use function interface_exists;
-use function is_dir;
-use function is_int;
-use function is_string;
-use function mkdir;
-use function rename;
-use function sprintf;
-use function str_replace;
-use function tempnam;
-use function unlink;
-
-use const PHP_EOL;
-use const PHP_VERSION_ID;
-
 /**
  * @template T of object
  */
 final class NullObject implements NullObjectInterface
 {
-    private const CLASS_TEMPLATE = <<<EOT
-%sclass %s implements \%s
-{
-%s}
-EOT;
-
     /**
      * {@inheritDoc}
      */
     public function newInstance(string $interface): object
     {
-        $generated = $this->generate($interface);
-        eval($generated->code);
-        $class = $generated->class;
-
-        /** @psalm-suppress MixedMethodCall */
-        $object = new $class();
-        assert($object instanceof $interface);
-
-        return $object;
+        return (new Code())->newInstance($interface);
     }
 
     /**
@@ -65,19 +22,7 @@ EOT;
      */
     public function generate(string $interface): GeneratedCode
     {
-        $class = new ReflectionClass($interface);
-        if (! interface_exists($class->getName())) {
-            throw new LogicException($class->getName());
-        }
-
-        $classMeta = $this->getClassMeta($class);
-
-        $shortNullClassName = $this->getNullClassShortName($class);
-        /** @var class-string $fqClassName */
-        $fqClassName = $this->getNullClassName($class);
-        $code = sprintf(self::CLASS_TEMPLATE, $classMeta, $shortNullClassName, $interface, $this->getMethods($class));
-
-        return new GeneratedCode($fqClassName, $code);
+        return (new Code())->generate($interface);
     }
 
     /**
@@ -85,134 +30,12 @@ EOT;
      */
     public function save(string $interface, string $scriptDir): string
     {
-        $nullClass = $this->getNullClassName(new ReflectionClass($interface));
-        $generated = $this->generate($interface);
-        $filePath = $generated->filePath($scriptDir);
-        $this->filePutContents($filePath, $generated->phpCode());
-        /** @psalm-suppress UnresolvableInclude */
-        if (! class_exists($nullClass, false)) {
-            require $filePath;
+        $nullFile = new NullObjectFile($interface, $scriptDir);
+        $maybeClassName = $nullFile->include();
+        if ($maybeClassName) {
+            return $maybeClassName;
         }
 
-        return $generated->class;
-    }
-
-    /**
-     * @param ReflectionClass<object> $class
-     */
-    private function getClassMeta(ReflectionClass $class): string
-    {
-        $file = (array) file((string) $class->getFileName());
-        $fileMeta = array_slice($file, 1, $class->getStartLine() - 2);
-
-        return implode('', $fileMeta);
-    }
-
-    /**
-     * @param ReflectionClass<object> $class
-     */
-    private function getMethods(ReflectionClass $class): string
-    {
-        $methods = $class->getMethods();
-        $methodStrings = [];
-        $file = (array) file((string) $class->getFileName());
-        foreach ($methods as $method) {
-            $interfaceMethod = implode(PHP_EOL, array_slice($file, $method->getStartLine() - 1, $method->getEndLine()  - $method->getStartLine() + 1));
-            $methodSignature = str_replace(';', "\n    {\n    }", $interfaceMethod);
-            $methodStrings[] = $this->getMethodMeta($method) . $methodSignature;
-        }
-
-        return implode(PHP_EOL, $methodStrings);
-    }
-
-    private function getMethodMeta(ReflectionMethod $method): string
-    {
-        $attr =  PHP_VERSION_ID >= 80000 ? sprintf("    %s\n", $this->getAttributes($method)) : '';
-
-        return sprintf("    %s\n%s", $method->getDocComment(), $attr);
-    }
-
-    /**
-     * @psalm-suppress MixedAssignment
-     */
-    private function getAttributes(ReflectionMethod $method): string
-    {
-        $attrs = $method->getAttributes();
-        $attrList = [];
-        if ($attrs) {
-            foreach ($attrs as $attr) {
-                /** @psalm-var ReflectionAttribute $attr */
-                $attrList[] = $this->getAttribute($attr); // @phpstan-ignore-line
-            }
-        }
-
-        return implode(PHP_EOL, $attrList);
-    }
-
-    public function filePutContents(string $filename, string $content): void
-    {
-        $dir = dirname($filename);
-        ! is_dir($dir) && ! mkdir($dir, 0777, true) && ! is_dir($dir);
-        $tmpFile = tempnam(dirname($filename), 'swap');
-        if (is_string($tmpFile) && file_put_contents($tmpFile, $content) && @rename($tmpFile, $filename)) {
-            return;
-        }
-
-        // @codeCoverageIgnoreStart
-        @unlink((string) $tmpFile);
-
-        throw new FileNotWritable($filename);
-        // @codeCoverageIgnoreEnd
-    }
-
-    /**
-     * @psalm-param ReflectionAttribute $attr
-     * @phpstan-param ReflectionAttribute<object> $attr
-     */
-    private function getAttribute(ReflectionAttribute $attr): string
-    {
-        /** @var array<float|int|string> $args */
-        $args = $attr->getArguments();
-        $argList = [];
-        foreach ($args as $key => $arg) {
-            $val = is_string($arg) ? sprintf("'%s'", $arg) : (string) $arg;
-            $argList[$key] = is_int($key) ? $val : "{$key}: $val";
-        }
-
-        /** @var class-string $class */
-        $class = $attr->getName();
-        $shortName = (new ReflectionClass($class))->getShortName();
-
-        return sprintf('#[%s(%s)]', $shortName, implode(', ', $argList));
-    }
-
-    /**
-     * @param ReflectionClass<object> $class
-     */
-    private function getNullClassName(ReflectionClass $class): string
-    {
-        return $class->getName() . $this->getTime($class) . 'aNull';
-    }
-
-    /**
-     * @param ReflectionClass<object> $class
-     */
-    private function getNullClassShortName(ReflectionClass $class): string
-    {
-        return $class->getShortName() . $this->getTime($class) . 'aNull';
-    }
-
-    /**
-     * @param ReflectionClass<object> $class
-     */
-    private function getTime(ReflectionClass $class): string
-    {
-        $time = 0;
-        while ($class instanceof ReflectionClass) {
-            $time .= filemtime($class->getFileName());
-            $class = $class->getParentClass();
-        }
-
-        return hash('crc32b', (string) $time);
+        return $nullFile->save();
     }
 }
