+++ tests/Fake/FakeChildInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Koriym\NullObject;
+
+interface FakeChildInterface
+{
+    public function doChild(): void;
+}
